# Use tracing appender to improve performance