Borrowing idea from [here](https://github.com/fermyon/ros/pull/4)

I ran some benchmarks against the default `rust-http` app before and after the changes.

### Before
```sh
$ bombardier localhost:3000
Bombarding http://localhost:3000 for 10s using 125 connection(s)
[==========================================================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     18513.31    1447.98   21170.52
  Latency        6.75ms     1.41ms    36.56ms
  HTTP codes:
    1xx - 0, 2xx - 185196, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     3.46MB/s
```

### After

```sh
$  bombardier localhost:3000
Bombarding http://localhost:3000 for 10s using 125 connection(s)
[==========================================================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     21480.65    1998.23   24816.04
  Latency        5.83ms     1.93ms    81.97ms
  HTTP codes:
    1xx - 0, 2xx - 214305, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     4.00MB/s
```

### Thoughts
We can see that there is a perf improvement, but I'm a bit confused as to why. Now that I think about it my test wasn't even producing any tracing (the thing this change is supposed to speed up) b/c I was doing `spin up` without `RUST_LOG`.

I'm curious if anyone has ideas why this created a speedup anyways.