@kate-goldenring 2024-06-25
@calebschoepp I'm curious about this too. Maybe shutdown is cleaner/faster with the guard