@calebschoepp 2024-06-25
> @calebschoepp I'm curious about this too. Maybe shutdown is cleaner/faster with the guard

Shutdown of Spin or do you mean cleanup of a Wasm instance? I don't see why Spin shutting down quicker would increase throughput.