@kate-goldenring 2024-06-25
> I don't see why Spin shutting down quicker would increase throughput.

Good point 🤷‍♀️