@calebschoepp 2024-06-25
Benchmarking with `RUST_LOG=info spin up`

### Before
```sh
$ Bombarding http://localhost:3000 for 10s using 125 connection(s)
[==========================================================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     17286.99    1759.51   20610.16
  Latency        7.23ms     1.71ms    53.75ms
  HTTP codes:
    1xx - 0, 2xx - 172968, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     3.23MB/s
```
### After
```sh
$ bombardier localhost:3000
Bombarding http://localhost:3000 for 10s using 125 connection(s)
[==========================================================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     16292.78    1933.98   22082.56
  Latency        7.67ms     2.15ms    60.37ms
  HTTP codes:
    1xx - 0, 2xx - 163007, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     3.04MB/s
```

### Thoughts

So it's worse after the change when tracing is on?! 🤷 I'm confused, if I can't come up with a reasonable explanation I'll probably just close this PR and leave it with the status quo.